requirements.txt
# AI DECLARANT MVP - Зависимости

# Веб-приложение
streamlit>=1.35.0

# Обработка данных
pandas>=2.1.0,<3.0.0
openpyxl>=3.1.2
xlsxwriter>=3.1.0
numpy>=1.24.0,<2.0.0

# Среда выполнения и конфигурация
python-dotenv>=1.0.0

# OpenAI Agents SDK (официальная библиотека)
openai-agents

# Валидация и типизация
pydantic>=2.0.0,<3.0.0

# Улучшенный интерфейс и визуализация
plotly>=5.15.0 